+++ src/components/TKMainComponent/TKCampComponents/TKCampToolbar.vue
@@ -5,7 +5,7 @@
         <v-autocomplete
           v-if="dataset.currentCamp"
           key="1"
-          class="tk-camp-toolbar-date"
+          class="tk-camp-toolbar-item"
           background-color="accent"
           color="background"
           :disabled="dataset.currentCamp.submissions.length < 2"
@@ -18,12 +18,12 @@
           :items="dataset.currentCamp.submissions"
           item-text="date"
           return-object
-          :prefix="$t('site.dateSuffix').toUpperCase()"
+          :prefix="$t('site.datePreffix').toUpperCase()"
         ></v-autocomplete>
         <v-autocomplete
           v-else
           key="2"
-          class="tk-camp-toolbar-date-disabled"
+          class="tk-camp-toolbar-item-disabled"
           background-color="accent"
           color="background"
           disabled
@@ -37,20 +37,13 @@
       </div>
     </transition>
 
-    <transition mode="out-in" name="fade-in">
-      <v-btn
-        :key="$root.$i18n.locale"
-        depressed
-        color="#000"
-        elevation="0"
-        class="tk-camp-toolbar-export"
-        height="44"
-        v-on:click="onExportTriggered"
-        :disabled="!dataset.currentCamp"
-      >
-        {{ $t("site.exportAsCSV") }}
-      </v-btn>
-    </transition>
+    <TKCampToolbarExportButton
+      :appConfig="appConfig"
+      :dataset="dataset"
+      :visualizerOptions="visualizerOptions"
+      :pdfInfos="pdfInfos"
+      class="tk-camp-toolbar-container"
+    />
 
     <v-menu
       :offset-y="true"
@@ -93,11 +86,18 @@
 </template>
 
 <script lang="ts">
-import { TKCSVWrite } from "@/domain/csv/TKCSVWriter";
 import { TKDatasetFilterer } from "@/domain/survey/TKDatasetFilterer";
 import { Component, Vue, Prop } from "vue-property-decorator";
 import { TKSubmissionVisualizerOptions } from "./TKSubmissionVisualizer";
-@Component
+import TKCampToolbarExportButton from "./TKCampToolbarExportButton.vue";
+import { TKOpsmapConfiguration } from "@/domain";
+import { TKPDFInfos } from "@/domain/survey/TKPDFInfos";
+
+@Component({
+  components: {
+    TKCampToolbarExportButton
+  }
+})
 export default class TKCampToolbar extends Vue {
   @Prop()
   readonly visualizerOptions!: TKSubmissionVisualizerOptions;
@@ -105,11 +105,11 @@ export default class TKCampToolbar extends Vue {
   @Prop()
   readonly dataset!: TKDatasetFilterer;
 
-  onExportTriggered() {
-    if (this.dataset && this.dataset.currentSubmission) {
-      TKCSVWrite(this.dataset, this.$root.$i18n.locale);
-    }
-  }
+  @Prop()
+  readonly appConfig!: TKOpsmapConfiguration;
+
+  @Prop()
+  readonly pdfInfos!: TKPDFInfos;
 }
 </script>
 
@@ -121,12 +121,8 @@ export default class TKCampToolbar extends Vue {
   align-items: top;
   column-gap: 5px;
 }
-.tk-camp-toolbar-date-container {
-  width: 40%;
-  min-width: 50px;
-}
 
-.tk-camp-toolbar-date-single {
+.tk-camp-toolbar-item.theme--light.v-input input {
   color: #fff !important;
   font-family: "Arial";
   font-weight: bold !important;
@@ -134,38 +130,15 @@ export default class TKCampToolbar extends Vue {
   letter-spacing: 0.86px !important;
 }
 
-.tk-camp-toolbar-export {
-  width: 40%;
-  min-width: 50px;
-  overflow: hidden;
-}
-.tk-camp-toolbar-export .v-btn__content {
-  color: #fff !important;
-  font-family: "Arial";
-  font-weight: bold !important;
-  font-size: 12px !important;
-  letter-spacing: 0.86px !important;
-}
-
-.tk-camp-toolbar-date.theme--dark.v-input input,
-.tk-camp-toolbar-date.theme--light.v-input input {
-  color: #fff !important;
-  font-family: "Arial";
-  font-weight: bold !important;
-  font-size: 12px !important;
-  letter-spacing: 0.86px !important;
-}
-
-.tk-camp-toolbar-date .v-icon.v-icon {
+.tk-camp-toolbar-item .v-icon.v-icon {
   color: #fff !important;
 }
 
-.tk-camp-toolbar-date .theme--dark.v-icon.v-icon.v-icon--disabled,
-.tk-camp-toolbar-date .theme--light.v-icon.v-icon.v-icon--disabled {
+.tk-camp-toolbar-item .theme--light.v-icon.v-icon.v-icon--disabled {
   opacity: 1 !important;
 }
 
-.tk-camp-toolbar-date-disabled.v-input--is-disabled .v-input__slot {
+.tk-camp-toolbar-item-disabled.v-input--is-disabled .v-input__slot {
   background-color: rgba(0, 0, 0, 0.12) !important;
 }