+++ src/components/TKMainComponent/TKMap/TKMap.vue
@@ -397,21 +397,21 @@ export default class TKMap extends Vue {
   position: absolute;
   top: 8px;
   right: 8px;
-  z-index: 2500;
+  z-index: 1;
 }
 
 .tk-basemap-picker {
   position: absolute;
   top: 8px;
   left: 8px;
-  z-index: 2500;
+  z-index: 1;
 }
 
 .tk-map-filters {
   position: absolute;
   bottom: 28px;
   right: 8px;
-  z-index: 2500;
+  z-index: 1;
 }
 </style>